# CLI: use one process to test many keys for mfkey32v2